Added some tests using mocks and stub
