+++ contacts.json
@@ -1,1 +1,4 @@
-{"name":"pepe","phone":8888,"_id":"Ps2frS8cSZURwyQE"}
+{"name":"juan","phone":5555,"_id":"5n4rY9jbTR4FDHY8"}
+{"name":"pepe","phone":6666,"_id":"92sl0rOat3eaBFFc"}
+{"name":"jaime","phone":1111,"_id":"Nzottc8cKjeQtDRD"}
+{"name":"jaime","phone":1111,"_id":"u4sFuDJX14pvybTr"}
