+++ package.json
@@ -50,6 +50,7 @@
     "karma-jasmine-html-reporter": "^0.2.2",
     "mocha": "^5.2.0",
     "protractor": "~5.4.0",
+    "sinon": "^7.1.1",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
     "typescript": "~3.1.6"
